requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.8
async-timeout==5.0.1
attrs==25.1.0
celery==5.3.4
certifi==2025.1.31
diff-match-patch==20241021
dill==0.3.9
Django==4.2.17
django-import-export==4.3.3
django-simple-history==3.7.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-silk==5.0.4
flake8==7.1.1
frozenlist==1.5.0
idna==3.10
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.1.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
redis==5.0.1
reportlab==4.1.0
sqlparse==0.5.3
tablib==3.7.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
yarl==1.18.3
sentry-sdk==1.40.4
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
dj-database-url==2.1.0
allauth==0.57.0
allauth-oauth2==0.1.0
